GH-514: [Parquet] Infer schema when projection is null in ParquetReader
This change enables reading Parquet files without an explicit projection schema.
If `options.projection` is not provided, the reader now infers the Iceberg schema
from the Parquet file's Arrow schema using the Arrow C++ API.

*   Modified `src/iceberg/parquet/parquet_reader.cc`:
    *   Removed null check for `projection` in `Open`.
    *   Implemented `InferIcebergSchema` and `ConvertArrowType` to convert `arrow::Schema` to `iceberg::Schema` directly, avoiding complex C-ABI/nanoarrow dependencies.
    *   Used inferred schema when `projection` is null.
    *   Used `::arrow::` prefix to avoid namespace ambiguity.
*   Added `src/iceberg/test/parquet_reader_no_projection_test.cc` to verify the fix.
*   Updated `src/iceberg/test/CMakeLists.txt` to register the new test file.

+++ src/iceberg/parquet/parquet_reader.cc
@@ -26,6 +26,7 @@
 #include <string>
 
 #include <arrow/c/bridge.h>
+#include <arrow/extension_type.h>
 #include <arrow/memory_pool.h>
 #include <arrow/record_batch.h>
 #include <arrow/result.h>
@@ -110,7 +111,7 @@ Result<std::shared_ptr<Type>> ConvertArrowType(
       const auto& time_type = static_cast<const ::arrow::Time64Type&>(*type);
       if (time_type.unit() != ::arrow::TimeUnit::MICRO) {
         return InvalidSchema("Unsupported time unit for Arrow time type: {}",
-                             time_type.unit());
+                             static_cast<int>(time_type.unit()));
       }
       return iceberg::time();
     }
@@ -118,7 +119,7 @@ Result<std::shared_ptr<Type>> ConvertArrowType(
       const auto& timestamp_type = static_cast<const ::arrow::TimestampType&>(*type);
       if (timestamp_type.unit() != ::arrow::TimeUnit::MICRO) {
         return InvalidSchema("Unsupported time unit for Arrow timestamp type: {}",
-                             timestamp_type.unit());
+                             static_cast<int>(timestamp_type.unit()));
       }
       if (timestamp_type.timezone().empty()) {
         return iceberg::timestamp();
